Add files via upload
MAUIcount is a Python script for the analysis of amplicon sequences that incorporate a random sequence tag (seqid or Unique Molecular Identifier).  The method is described in the following preprint: https://www.biorxiv.org/content/10.1101/538587v2.abstract.
MC_parameters is a file that is edited by the user to set parameters for MAUIcount.